:recycle: Refactor: Strategies
