+++ zerohertzLib/quant/__init__.py
@@ -12,7 +12,7 @@
 """
 
 from zerohertzLib.quant.backtest import Experiments, backtest, experiments
-from zerohertzLib.quant.strategy import bollinger_bands, momentum, moving_average, rsi
+from zerohertzLib.quant.strategies import bollinger_bands, momentum, moving_average, rsi
 
 __all__ = [
     "moving_average",
